getvite_api/requirements.txt
flask
sqlalchemy
sqlalchemy-cockroachdb
psycopg2
flask-cors
